requirements.txt
et-xmlfile==1.1.0
Django==3.2.11
django-cors-headers==3.13.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2 
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.2.4
PyPDF2==2.1.0
